requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-sms==0.5.0
django-taggit==1.4.0
django-taggit-serializer==0.1.7
djangorestframework==3.12.4
idna==2.10
mysqlclient==2.0.3
oauthlib==3.1.1
Pillow==8.2.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
